Updated to return false instead of "Invalid Response"
Invalid responses and "non-threatening" IP address results come back as the same thing. I couldn't find a way to get it to return something more alerting when the DNS lookup failed, so this will have to do. If you're caching your results for a short period of time, then any issues with the API should correct themselves within the TTL of your data